Make `jsanitize(recursive_msonable=True)` respect duck typing
"If it looks like a duck and quacks like a duck..."

In this PR, I replace the `isinstance(obj, MSONable)` check with a `hasattr(obj, "as_dict")` check instead in `jsanitize`. This makes it possible to jsanitize objects that have been patched with `.as_dict()` attributes even if they are not formally inheriting from the `MSONable` class.